Split stationarity and Cointegration
