+++ mlfinlab/statistical_arbitrage/__init__.py
@@ -2,4 +2,4 @@
 Module for Statistical Arbitrage.
 """
 
-from mlfinlab.statistical_arbitrage.check_conditions import calc_stationarity
+from mlfinlab.statistical_arbitrage.stationarity import calc_stationarity
